Реализована логика сохранения и обновления КП

+++ server/src/main/kotlin/com/omsk/kp/domain/model/CommercialOfferDetails.kt
@@ -10,12 +10,12 @@ import java.time.Instant
 data class CommercialOfferDetails(
     val name: String,
     val price: Int,
-    val markupExtra: Int? = null,
-    val markupPercent: Int? = null,
-    val transportExtra: Int? = null,
-    val transportPercent: Int? = null,
+    val markupExtra: Double? = null,
+    val markupPercent: Double? = null,
+    val transportExtra: Double? = null,
+    val transportPercent: Double? = null,
     val quantity: Int,
-    val weightKg: Int,
+    val weightKg: Double,
     val commercialOfferId: Long,
     val createdAt: Instant = Instant.now(),
